NFDIV-2373: Action dev review comments
