+++ src/test/java/uk/gov/hmcts/divorce/citizen/event/CitizenSwitchedToSoleCoTest.java
@@ -11,24 +11,14 @@ import uk.gov.hmcts.ccd.sdk.api.Event;
 import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
 import uk.gov.hmcts.divorce.citizen.notification.Applicant1SwitchToSoleCoNotification;
 import uk.gov.hmcts.divorce.citizen.notification.Applicant2SwitchToSoleCoNotification;
-import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
+import uk.gov.hmcts.divorce.citizen.service.SwitchToSoleService;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.ConditionalOrder;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.divorcecase.model.UserRole;
-import uk.gov.hmcts.divorce.idam.IdamService;
 import uk.gov.hmcts.divorce.notification.NotificationDispatcher;
 import uk.gov.hmcts.divorce.solicitor.service.CcdAccessService;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.ccd.client.CaseAssignmentApi;
-import uk.gov.hmcts.reform.ccd.client.model.CaseAssignmentUserRole;
-import uk.gov.hmcts.reform.ccd.client.model.CaseAssignmentUserRoleWithOrganisation;
-import uk.gov.hmcts.reform.ccd.client.model.CaseAssignmentUserRolesRequest;
-import uk.gov.hmcts.reform.ccd.client.model.CaseAssignmentUserRolesResource;
-import uk.gov.hmcts.reform.idam.client.models.User;
-import uk.gov.hmcts.reform.idam.client.models.UserDetails;
 
-import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -42,12 +32,8 @@ import static uk.gov.hmcts.divorce.citizen.event.CitizenSwitchedToSoleCo.SWITCH_
 import static uk.gov.hmcts.divorce.divorcecase.model.ApplicationType.SOLE_APPLICATION;
 import static uk.gov.hmcts.divorce.divorcecase.model.ConditionalOrder.D84WhoApplying.APPLICANT_1;
 import static uk.gov.hmcts.divorce.divorcecase.model.ConditionalOrder.D84WhoApplying.APPLICANT_2;
-import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.CREATOR;
 import static uk.gov.hmcts.divorce.testutil.ConfigTestUtil.createCaseDataConfigBuilder;
 import static uk.gov.hmcts.divorce.testutil.ConfigTestUtil.getEventsFrom;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.CASEWORKER_AUTH_TOKEN;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.CASEWORKER_USER_ID;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_SERVICE_AUTH_TOKEN;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.validJointApplicant1CaseData;
 
 @ExtendWith(MockitoExtension.class)
@@ -69,13 +55,7 @@ class CitizenSwitchedToSoleCoTest {
     private NotificationDispatcher notificationDispatcher;
 
     @Mock
-    private CaseAssignmentApi caseAssignmentApi;
-
-    @Mock
-    private IdamService idamService;
-
-    @Mock
-    private AuthTokenGenerator authTokenGenerator;
+    private SwitchToSoleService switchToSoleService;
 
     @InjectMocks
     private CitizenSwitchedToSoleCo citizenSwitchedToSoleCo;
@@ -133,7 +113,7 @@ class CitizenSwitchedToSoleCoTest {
     }
 
     @Test
-    void shouldSwitchUserDataIfApplicant2TriggeredD84SwitchToSole() {
+    void shouldSwitchUserDataAndRolesIfApplicant2TriggeredD84SwitchToSole() {
         final long caseId = 1L;
         CaseData caseData = validJointApplicant1CaseData();
         caseData.setConditionalOrder(ConditionalOrder.builder().d84WhoApplying(APPLICANT_2).build());
@@ -141,97 +121,11 @@ class CitizenSwitchedToSoleCoTest {
             .id(caseId)
             .data(caseData)
             .build();
-
-        final CaseAssignmentUserRolesResource caseRolesResponse = CaseAssignmentUserRolesResource.builder()
-            .caseAssignmentUserRoles(List.of(
-                CaseAssignmentUserRole.builder().userId("1").caseRole("[APPLICANTTWO]").build(),
-                CaseAssignmentUserRole.builder().userId("2").caseRole("[CREATOR]").build()
-            ))
-            .build();
-
-        final UserDetails userDetails = UserDetails.builder().id(CASEWORKER_USER_ID).build();
-        final User user = new User(CASEWORKER_AUTH_TOKEN, userDetails);
-
-        when(idamService.retrieveSystemUpdateUserDetails()).thenReturn(user);
-        when(authTokenGenerator.generate()).thenReturn(TEST_SERVICE_AUTH_TOKEN);
-        when(caseAssignmentApi.getUserRoles(CASEWORKER_AUTH_TOKEN, TEST_SERVICE_AUTH_TOKEN, List.of(String.valueOf(caseId))))
-            .thenReturn(caseRolesResponse);
-
-        final Applicant applicant1BeforeSwitch = caseData.getApplicant1();
-        final Applicant applicant2BeforeSwitch = caseData.getApplicant2();
-
-        final AboutToStartOrSubmitResponse<CaseData, State> response = citizenSwitchedToSoleCo.aboutToSubmit(caseDetails, caseDetails);
-
-        assertThat(response.getData().getApplicant1()).isEqualTo(applicant2BeforeSwitch);
-        assertThat(response.getData().getApplicant2()).isEqualTo(applicant1BeforeSwitch);
-    }
-
-    @Test
-    void shouldSwitchUserRolesIfApplicant2TriggeredD84SwitchToSole() {
-        final long caseId = 1L;
-        CaseData caseData = validJointApplicant1CaseData();
-        caseData.setConditionalOrder(ConditionalOrder.builder().d84WhoApplying(APPLICANT_2).build());
-        final CaseDetails<CaseData, State> caseDetails = CaseDetails.<CaseData, State>builder()
-            .id(caseId)
-            .data(caseData)
-            .build();
-
-        final CaseAssignmentUserRolesResource caseRolesResponse = CaseAssignmentUserRolesResource.builder()
-            .caseAssignmentUserRoles(List.of(
-                CaseAssignmentUserRole.builder().userId("1").caseRole("[APPLICANTTWO]").build(),
-                CaseAssignmentUserRole.builder().userId("2").caseRole("[CREATOR]").build()
-            ))
-            .build();
-
-        final UserDetails userDetails = UserDetails.builder().id(CASEWORKER_USER_ID).build();
-        final User user = new User(CASEWORKER_AUTH_TOKEN, userDetails);
-
-        when(idamService.retrieveSystemUpdateUserDetails()).thenReturn(user);
-        when(authTokenGenerator.generate()).thenReturn(TEST_SERVICE_AUTH_TOKEN);
-        when(caseAssignmentApi.getUserRoles(CASEWORKER_AUTH_TOKEN, TEST_SERVICE_AUTH_TOKEN, List.of(String.valueOf(caseId))))
-            .thenReturn(caseRolesResponse);
-        when(ccdAccessService.getCaseAssignmentRequest(caseId, "2", null, CREATOR))
-            .thenReturn(getCaseAssignmentRequest("2", CREATOR));
-        when(ccdAccessService.getCaseAssignmentRequest(caseId, "1", null, UserRole.APPLICANT_2))
-            .thenReturn(getCaseAssignmentRequest("1", UserRole.APPLICANT_2));
-        when(ccdAccessService.getCaseAssignmentRequest(caseId, "1", null, CREATOR))
-            .thenReturn(getCaseAssignmentRequest("1", CREATOR));
-        when(ccdAccessService.getCaseAssignmentRequest(caseId, "2", null, UserRole.APPLICANT_2))
-            .thenReturn(getCaseAssignmentRequest("2", UserRole.APPLICANT_2));
 
         citizenSwitchedToSoleCo.aboutToSubmit(caseDetails, caseDetails);
 
-        verify(caseAssignmentApi)
-            .getUserRoles(
-                CASEWORKER_AUTH_TOKEN,
-                TEST_SERVICE_AUTH_TOKEN,
-                List.of(String.valueOf(caseId))
-            );
-        verify(caseAssignmentApi)
-            .removeCaseUserRoles(
-                CASEWORKER_AUTH_TOKEN,
-                TEST_SERVICE_AUTH_TOKEN,
-                getCaseAssignmentRequest("2", CREATOR)
-            );
-        verify(caseAssignmentApi)
-            .removeCaseUserRoles(
-                CASEWORKER_AUTH_TOKEN,
-                TEST_SERVICE_AUTH_TOKEN,
-                getCaseAssignmentRequest("1", UserRole.APPLICANT_2)
-            );
-        verify(caseAssignmentApi)
-            .addCaseUserRoles(
-                CASEWORKER_AUTH_TOKEN,
-                TEST_SERVICE_AUTH_TOKEN,
-                getCaseAssignmentRequest("1", CREATOR)
-            );
-        verify(caseAssignmentApi)
-            .addCaseUserRoles(
-                CASEWORKER_AUTH_TOKEN,
-                TEST_SERVICE_AUTH_TOKEN,
-                getCaseAssignmentRequest("2", UserRole.APPLICANT_2)
-            );
-        verifyNoMoreInteractions(caseAssignmentApi);
+        verify(switchToSoleService).switchUserRoles(caseId);
+        verify(switchToSoleService).switchApplicantData(caseData);
     }
 
     @Test
@@ -246,24 +140,6 @@ class CitizenSwitchedToSoleCoTest {
 
         citizenSwitchedToSoleCo.aboutToSubmit(caseDetails, caseDetails);
 
-        verifyNoInteractions(idamService);
-        verifyNoInteractions(authTokenGenerator);
-        verifyNoInteractions(caseAssignmentApi);
-    }
-
-    private CaseAssignmentUserRolesRequest getCaseAssignmentRequest(String userId, UserRole role) {
-        return CaseAssignmentUserRolesRequest.builder()
-            .caseAssignmentUserRolesWithOrganisation(
-                List.of(getCaseAssignmentUserRole(role.getRole(), userId))
-            ).build();
-    }
-
-    private CaseAssignmentUserRoleWithOrganisation getCaseAssignmentUserRole(String role, String userId) {
-        return CaseAssignmentUserRoleWithOrganisation.builder()
-            .organisationId(null)
-            .caseDataId(String.valueOf(1L))
-            .caseRole(role)
-            .userId(userId)
-            .build();
+        verifyNoInteractions(switchToSoleService);
     }
 }
